Update senseHat_MemoryGame.py
Defined the initial code as a function added two others for the initialization of the game at start. 

load_character() reads the CSV file and in one line procures the character for display through 64 separate letters (a and b both representing tuples for the LED display), as well as assigning that character's true meaning and pronunciation from the same row (after the 64 letters). 

create_meaning_list() then creates a random list of four items, one of which is the true meaning and the other three are incorrect meanings drawn from the list extra_meaning.


create_pronunciation_list() then creates a random list of four items, one of which is the true pronunciation and the other three are incorrect meanings drawn from the list extra_pronunciation.